Add possibility to specify Home Assistant device class for more devices

+++ src/main/kotlin/com/mqgateway/core/io/JCommSerial.kt
@@ -33,7 +33,7 @@ class JCommSerial : Serial {
 
   override fun write(message: String) {
     val bytes = message.toByteArray()
-    serialPort!!.writeBytes(bytes, bytes.size.toLong())
+    serialPort!!.writeBytes(bytes, bytes.size)
   }
 }
